+++ NEWS.md
@@ -3,4 +3,5 @@
 ## New functions
 
 * `merge_df` to fully join labelled data frame and preserve value and variable labels.
-
+* `wtd_sd` to compute weighted standard deviations.
+* `wtd_se` to compute weighted standard errors.